backend/requirements.txt
# LangChain + DeepSeek API
langchain>=0.2.12
langchain-community>=0.2.10
langchain-openai>=0.1.8
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.2.0

# Web API
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9

# 微信公众号发布
wechatpy>=1.13.0
google-search-results>=2.4.0

fpdf>=1.7.2